Move all corner translates into same file
