+++ tests/translate/__init__.py
@@ -16,6 +16,7 @@ from .translate_corners import (
     TranslateCopyCorners,
     TranslateFillCorners2D,
     TranslateFillCornersCells,
+    TranslateFillCornersVector,
 )
 from .translate_cubedtolatlon import TranslateCubedToLatLon
 from .translate_d2a2c_vect import TranslateD2A2C_Vect
@@ -25,9 +26,6 @@ from .translate_del6vtflux import TranslateDel6VtFlux
 from .translate_delnflux import TranslateDelnFlux, TranslateDelnFlux_2
 from .translate_divergencedamping import TranslateDivergenceDamping
 from .translate_dyncore import TranslateDynCore
-from .translate_fill4corners import TranslateFill4Corners
-from .translate_fillcorners import TranslateFillCorners
-from .translate_fillcornersvector import TranslateFillCornersVector
 from .translate_fillz import TranslateFillz
 from .translate_fluxcapacitor import TranslateFluxCapacitor
 from .translate_fvsetup import TranslateFVSetup
