+++ tests/translate/translate_corners.py
@@ -66,3 +66,26 @@ class TranslateCopyCorners(TranslateFortranData2Py):
             raise ValueError("Invalid input")
         corners.copy_corners(inputs["q"], direction, self.grid)
         return {"q": inputs["q"]}
+
+
+class TranslateFillCornersVector(TranslateFortranData2Py):
+    def __init__(self, grid):
+        super().__init__(grid)
+        self.in_vars["data_vars"] = {"vc": {}, "uc": {}, "nord_col": {}}
+        self.out_vars = {"vc": grid.y3d_domain_dict(), "uc": grid.x3d_domain_dict()}
+
+    def compute(self, inputs):
+        nord_column = inputs["nord_col"][0, 0, :]
+        vector = True
+        self.make_storage_data_input_vars(inputs)
+        num_k = self.grid.npz
+        for nord in np.unique(nord_column):
+            if nord != 0:
+                ki = [i for i in range(num_k) if nord_column[i] == nord]
+                d = utils.k_slice(inputs, ki)
+                self.grid.npz = len(ki)
+                corners.fill_corners_dgrid(d["vc"], d["uc"], self.grid, vector)
+                inputs["vc"][:, :, ki] = d["vc"]
+                inputs["uc"][:, :, ki] = d["uc"]
+        self.grid.npz = num_k
+        return self.slice_output(inputs)
